# spec/openapi.yaml
swagger: '2.0'
info:
  title: AI Gateway
  version: 1.0.0
  description: |
    A platform for machine learning & high performance computing
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
host: 'localhost:9000'
basePath: /api/v1
securityDefinitions:
  api-authorizer:
    name: Authorization
    type: apiKey
    in: header
paths:
  '/configurations':
    get:
      description: |
        returns app configurations
      operationId: getConfigurations
      tags:
        - app
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Configuration'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        set app configurations
      operationId: postConfigurations
      tags:
        - app
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: AccountInfo
            type: object
            properties:
              docker_registry:
                description: Docker Registry endpoint
                type: string
              docker_hostname:
                description: Hostname for the private Docker registry
                type: string
              docker_username:
                description: Username for the private Docker registry
                type: string
              docker_password:
                description: Password for the private Docker registry
                type: string
                format: password
              ngc_email:
                description: E-mail address for NGC console
                type: string
                format: email
              ngc_password:
                description: Password for NGC console
                type: string
                format: password
              ngc_apikey:
                description: NGC - API Key
                type: string
              k8s_config:
                description: kubecfg
                type: string
              rescale_platform:
                description: Rescale platform endopoint
                type: string
                enum:
                  - https://platform.rescale.com
                  - https://platform.rescale.jp
                  - https://kr.rescale.com
                  - https://eu.rescale.com
              rescale_key:
                description: Rescale - API Key
                type: string
      responses:
        '201':
          description: OK
          schema:
            $ref: '#/definitions/Session'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/images':
    get:
      description: |
        returns local images
      operationId: getImages
      tags:
        - image
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Image'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        pull a specified image from Docker registry
      operationId: postNewImage
      tags:
        - image
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: ImageName
            type: object
            required:
              - image
            properties:
              image:
                description: Docker image name
                type: string
      responses:
        '201':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: |
        delete a specified local image
      operationId: deleteImage
      tags:
        - image
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: ImageName
            type: object
            required:
              - image
            properties:
              image:
                description: Docker image name
                type: string
      responses:
        '204':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Configuration:
    description: app's configurations
    required:
      - must_signed_in
    properties:
      must_signed_in:
        description: Users should be signed in
        type: string
        enum:
          - "yes"
          - "no"
      use_private_registry:
        description: Private registry will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_ngc:
        description: NGC will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_k8s:
        description: Kubernetes will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_rescale:
        description: Rescale will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      docker_registry:
        description: Docker Registry endpoint
        type: string
      docker_hostname:
        description: Hostname for the private Docker registry
        type: string
      docker_username:
        description: Username for the private Docker registry
        type: string
      docker_password:
        description: Fist 3 chars of the password of the private Docker registry
        type: string
      ngc_email:
        description: E-mail address for NGC console
        type: string
        format: email
      ngc_password:
        description: Fist 3 chars of the password for NGC console
        type: string
      ngc_apikey:
        description: Fist 5 chars of NGC API Key
        type: string
      k8s_config:
        description: kubecfg
        type: string
      rescale_platform:
        type: string
        enum:
          - https://platform.rescale.com
          - https://platform.rescale.jp
          - https://kr.rescale.com
          - https://eu.rescale.com
      rescale_key:
        description: Fist 5 chars of Rescal API Key
        type: string
  Session:
    description: docker image information
    required:
      - token
    properties:
      token:
        description: API Token
        type: string
...
  Image:
    description: docker image information
    required:
      - id
      - repoTags
    properties:
      id:
        description: the image ID
        type: string
      parentID:
        description: the image parent ID
        type: string
      repoDigests:
        description: the image digests
        type: array
        items:
          type: string
      repoTags:
        description: the image name/tag
        type: array
        items:
          type: string
      status:
        description: status of the image
        type: string
      size:
        description: size of the image
        type: integer
        format: int64
      virtualSize:
        description: virtual size of the image
        type: integer
        format: int64
      created:
        description: created unix timestamp
        type: string
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: string
      message:
        type: string